First/requirements.txt
langchain-community
langchain-openai
langchain
python-dotenv
langchain-google-genai
langchain_scrapegraphai